requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
boto3==1.24.58
botocore==1.27.58
certifi==2022.6.15
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.3
jmespath==1.0.1
multidict==6.0.2
pyTelegramBotAPI==4.7.0
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
yarl==1.8.1
